requirements.txt
Django
djangorestframework
markdown
django-filter
ConfigParser
mysqlclient
requests
feedparser
Gunicorn
whitenoise
django-heroku
python-decouple
psycopg2-binary
social-auth-app-django